fix binary output

+++ src/bdm/device/gpio/gpio_wrapper.cxx
@@ -11,18 +11,18 @@ namespace bdm::device::gpio {
 	}
 
 	void gpio_wrapper::set_pin_mode(std::uint8_t pin, io_mode m) {
-		LOG(DEBUG) << "gpio.set_pin_mode: pin=" << pin << ", mode=" << static_cast<int>(m);
+		LOG(DEBUG) << "gpio.set_pin_mode: pin=" << static_cast<int>(pin) << ", mode=" << static_cast<int>(m);
 		pinMode(pin, static_cast<int>(m));
 	}
 
 	void gpio_wrapper::write(std::uint8_t pin, bool high) {
-		LOG(DEBUG) << "gpio.write: pin=" << pin << ", high=" << std::boolalpha << high;
+		LOG(DEBUG) << "gpio.write: pin=" << static_cast<int>(pin) << ", high=" << std::boolalpha << high;
 		digitalWrite(pin, high);
 	}
 
 	bool gpio_wrapper::read(std::uint8_t pin) {
 		int val = digitalRead(pin);
-		LOG(DEBUG) << "gpio.read: pin=" << pin << ", high=" << val;
+		LOG(DEBUG) << "gpio.read: pin=" << static_cast<int>(pin) << ", high=" << val;
 		return val != 0;
 	}
 }
